#8108 dashboard contact statistics
